% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_table_name}
\alias{validate_table_name}
\title{Validate table name}
\usage{
validate_table_name(name)
}
\description{
Handles string, \code{I()} or \code{Id} objects and returns a string
}
\keyword{internal}
