% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workers.R
\name{write_rows}
\alias{write_rows}
\title{write_rows}
\usage{
write_rows(
  source,
  dest_table_name,
  con,
  mode = c("append", "insert", "update", "upsert", "patch", "delete"),
  ...
)
}
\description{
write_rows
}
\keyword{internal}
