% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer-package.R
\docType{package}
\name{writer-package}
\alias{writer}
\alias{writer-package}
\title{writer: Write from Multiple Sources to a Database Table}
\description{
Provides unified syntax to write data from lazy 'dplyr' 'tbl' or 'dplyr' 'sql' query or a dataframe to a database table with modes such as create, append, insert, update, upsert, patch, delete, overwrite, overwrite_schema.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/talegari/writer}
}

}
\author{
\strong{Maintainer}: Komala Sheshachala Srikanth \email{sri.teach@gmail.com} (\href{https://orcid.org/0000-0002-1865-5668}{ORCID})

}
\keyword{internal}
